refactor(contracts)!: Consolidate VotesERC20LockableV1 into VotesERC20V1
This commit introduces a major architectural simplification by merging the `VotesERC20LockableV1` contract's functionality directly into `VotesERC20V1`. This consolidation eliminates a redundant contract, reduces code duplication, and creates a single, more flexible and powerful base ERC20 token contract for the project.

The `VotesERC20V1` now includes all features previously found in the lockable version, such as token locking, maximum supply caps, and role-based minting. As part of this change, `VotesERC20V1` has been upgraded from `Ownable2StepUpgradeable` to `AccessControlUpgradeable` to support the more granular permissions required by these new features.

The `SystemDeployerV1` has been significantly updated to reflect this new, streamlined architecture, simplifying its token deployment logic and API.

**BREAKING CHANGE:**
- The `VotesERC20LockableV1` contract, its interface, and its test file have been deleted.
- `VotesERC20V1` now uses `AccessControlUpgradeable` instead of `Ownable2StepUpgradeable`. The `owner()` function is replaced by `hasRole(DEFAULT_ADMIN_ROLE, ...)`.
- The `initialize` function on `IVotesERC20V1` now includes `locked_` and `maxTotalSupply_` parameters.
- The `SystemDeployerV1.setupSafe` function signature has changed. It now accepts a simple array of `VotesERC20V1Params[]` instead of a complex struct containing two separate token type arrays.
- The `TypeTokenIndex` struct has been removed from `ISystemDeployerV1`. Adapters pointing to newly created tokens now use a simple `newTokenIndex` to identify the token in the `VotesERC20V1Params[]` array.

### Change Summary:
- **Contract Consolidation**: Merged all features from `VotesERC20LockableV1` (locking, supply caps, minting) into `VotesERC20V1`.
- **Access Control**: Upgraded `VotesERC20V1` from `Ownable` to `AccessControl` for finer-grained permission management.
- **Code Deletion**: Removed `VotesERC20LockableV1.sol`, `IVotesERC20LockableV1.sol`, and `VotesERC20LockableV1.test.ts`.
- **SystemDeployer Refactor**: Simplified the `SystemDeployerV1` contract and its tests, removing all logic related to the deprecated `VotesERC20LockableV1`.
- **Test Consolidation**: Merged all relevant tests from `VotesERC20LockableV1.test.ts` into `VotesERC20V1.test.ts` to cover the unified functionality.
