+++ contracts/interfaces/decent/deployables/IVotesERC20V1.sol
@@ -2,6 +2,17 @@
 pragma solidity ^0.8.30;
 
 interface IVotesERC20V1 {
+    // --- Errors ---
+
+    error IsLocked();
+    error ExceedMaxTotalSupply();
+    error InvalidMaxTotalSupply();
+
+    // --- Events ---
+
+    event Locked(bool isLocked);
+    event MaxTotalSupplyUpdated(uint256 newMaxTotalSupply);
+
     // --- Structs ---
 
     struct Metadata {
@@ -19,7 +30,9 @@ interface IVotesERC20V1 {
     function initialize(
         Metadata calldata metadata_,
         Allocation[] calldata allocations_,
-        address owner_
+        address owner_,
+        bool locked_,
+        uint256 maxTotalSupply_
     ) external;
 
     // --- Pure Functions ---
@@ -29,4 +42,20 @@ interface IVotesERC20V1 {
     // --- View Functions ---
 
     function clock() external view returns (uint48 clock);
+
+    function locked() external view returns (bool isLocked);
+
+    function maxTotalSupply() external view returns (uint256 maxTotalSupply);
+
+    function getUnlockTime() external view returns (uint48 unlockTime);
+
+    // --- State-Changing Functions ---
+
+    function lock(bool locked_) external;
+
+    function setMaxTotalSupply(uint256 newMaxTotalSupply_) external;
+
+    function mint(address to_, uint256 amount_) external;
+
+    function burn(uint256 amount_) external;
 }
